requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-environ==0.4.5
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.11.0
djangorestframework-api-key==1.4.1
get==2019.4.13
idna==2.9
logger==1.4
post==2019.4.13
psycopg2==2.8.4
public==2019.4.13
pytz==2019.3
query-string==2019.4.13
redis==3.4.1
request==2019.4.13
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
